backend/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
Flask-JWT-Extended>=4.6.0
Flask-CORS>=4.0.0
Flask-Mail>=0.10
Flask-Limiter>=3.5.0
marshmallow>=3.20.1
marshmallow-sqlalchemy>=0.30.0
PyMySQL>=1.1.0
cryptography>=41.0.7
python-dotenv>=1.0.0
Werkzeug>=3.0.1
razorpay>=1.4.1
celery>=5.3.6
redis>=5.0.1
Pillow>=10.2.0
openpyxl>=3.1.2
msoffcrypto-tool>=6.0.0
reportlab>=4.0
